[Add and Change] The shopinventory is complete and can be started with a script.

+++ src/main/java/net/firiz/renewatelier/alchemy/material/Category.java
@@ -45,32 +45,32 @@ public enum Category {
     STRING("糸素材", Material.STRING),
     POTION_MATERIAL("薬の材料", Material.SPIDER_EYE),
     POISON("毒の材料", Material.POISONOUS_POTATO),
-    MAGIC_ITEM("魔法の道具", new DoubleData<>(Material.DIAMOND_AXE, (short) 1524)),
-    MYSTERY("神秘の力", new DoubleData<>(Material.DIAMOND_AXE, (short) 0)), // 竜核画像
-    AN_ELIXIR("エリキシル", new DoubleData<>(Material.DIAMOND_AXE, (short) 0)), // 竜のウロコ画像
-    NEUTRALIZATION("中和剤", new DoubleData<>(Material.POTION, (short) 0)), // ポーション色あり
-    CLOTH("布", new DoubleData<>(Material.DIAMOND_AXE, (short) 0)), // 布画像
-    TNT("爆弾", new DoubleData<>(Material.DIAMOND_AXE, (short) 0)), // フラム画像
+    MAGIC_ITEM("魔法の道具", new DoubleData<>(Material.DIAMOND_AXE, 1524)),
+    MYSTERY("神秘の力", new DoubleData<>(Material.DIAMOND_AXE, 0)), // 竜核画像
+    AN_ELIXIR("エリキシル", new DoubleData<>(Material.DIAMOND_AXE, 0)), // 竜のウロコ画像
+    NEUTRALIZATION("中和剤", new DoubleData<>(Material.POTION, 0)), // ポーション色あり
+    CLOTH("布", new DoubleData<>(Material.DIAMOND_AXE, 0)), // 布画像
+    TNT("爆弾", new DoubleData<>(Material.DIAMOND_AXE, 0)), // フラム画像
     INGOT("金属", Material.IRON_INGOT),
     SWEETS("お菓子", Material.CAKE),
     FOOD("食品", Material.BREAD),
     WEAPON_MATERIAL("武器素材", Material.GOLD_INGOT),
-    ARMOR_MATERIAL("防具素材", new DoubleData<>(Material.DIAMOND_AXE, (short) 0)),// 布画像
+    ARMOR_MATERIAL("防具素材", new DoubleData<>(Material.DIAMOND_AXE, 0)),// 布画像
     WEAPON("武器", Material.IRON_SWORD),
     ARMOR("防具", Material.IRON_CHESTPLATE),
-    DECORATION("装飾品", new DoubleData<>(Material.DIAMOND_AXE, (short) 0)), // イヤリング画像
+    DECORATION("装飾品", new DoubleData<>(Material.DIAMOND_AXE, 0)), // イヤリング画像
     FLOWER("花", Material.DANDELION),
     SAND("砂", Material.SAND),
-    PUNIPUNI("プニプニ玉", new DoubleData<>(Material.DIAMOND_AXE, (short) 0)), // ぷにぷに玉画像
+    PUNIPUNI("プニプニ玉", new DoubleData<>(Material.DIAMOND_AXE, 0)), // ぷにぷに玉画像
     FISH("魚介類", Material.PUFFERFISH),
     OBJECT("オブジェ", Material.CAULDRON),
     IMPORTANT("重要", Material.FILLED_MAP),
-    CATALYST("触媒", new DoubleData<>(Material.DIAMOND_AXE, (short) 0)), // 触媒画像
+    CATALYST("触媒", new DoubleData<>(Material.DIAMOND_AXE, 0)), // 触媒画像
     COLLECTION_TOOL("採取道具", Material.IRON_PICKAXE);
 
     private static final Map<String, Category> BY_NAME = Maps.newHashMap();
     private final String name;
-    private final DoubleData<Material, Short> material;
+    private final DoubleData<Material, Integer> material;
 
     static {
         Category[] var3;
@@ -84,10 +84,10 @@ public enum Category {
 
     Category(String name, Material material) {
         this.name = name;
-        this.material = new DoubleData<>(material, (short) 0);
+        this.material = new DoubleData<>(material, 0);
     }
 
-    Category(String name, DoubleData<Material, Short> material) {
+    Category(String name, DoubleData<Material, Integer> material) {
         this.name = name;
         this.material = material;
     }
@@ -96,7 +96,7 @@ public enum Category {
         return "(".concat(name).concat(")");
     }
 
-    public DoubleData<Material, Short> getMaterial() {
+    public DoubleData<Material, Integer> getMaterial() {
         return material;
     }
